Add very simple echo server
This is a trivial echo server that sends the first 1024 bytes to the
sender. This first pass is a single-threaded, blocking server that
can't respond correctly to any HTTP requests. It handles all errors,
though not necessarily in the best way, and there is no logging
functionality, syslog or otherwise